Add multiple packages to organise classes
