+++ src/main/java/shagbot/exceptions/ShagBotException.java
@@ -1,4 +1,4 @@
-package Shagbot;
+package shagbot.exceptions;
 
 public class ShagBotException extends Exception{
