+++ src/main/java/shagbot/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package Shagbot;
+package shagbot.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
