+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
      ____________________________________________________________
-     Hello! I'm Shagbot
+     Hello! I'm shagbot
      What can I do for you?
      ____________________________________________________________
      ____________________________________________________________
